refactor(gpu_execution_prover): split tracing.rs into by-concern submodules
Pure code relocation: gpu/execution_prover/src/tracing.rs (886 lines, 5
distinct concerns) becomes a directory module with tracing/{ty,ranges,tracer,
producer,producers}.rs, keeping every crate::tracing::X path stable via a
flat glob re-export in tracing/mod.rs. No logic changes; only the module
plumbing plus the minimum pub(crate) markers needed for TracingDataProducerType
and TracingDataProducer to compile across the new submodule boundary (they
were previously private within the single file). The impl_write_delegation!
macro stays co-located with its two invocation sites in tracer.rs.
