--- gpu/execution_prover/src/tracing.rs
@@ -1,886 +1,0 @@
-use crate::messages::{TracingData, WorkerResult};
-use crate::A;
-use crossbeam_channel::{Receiver, Sender};
-use gpu_circuit_prover::prover::trace::tracing_data::{
-    DelegationTracingDataHostSource, TracingDataHost, UnrolledTracingDataHost,
-};
-use gpu_circuit_prover::witness::circuit_type::{
-    CircuitType, DelegationCircuitType, UnrolledCircuitType, UnrolledMemoryCircuitType,
-    UnrolledNonMemoryCircuitType,
-};
-use gpu_circuit_prover::witness::trace::ChunkedTraceHolder;
-use gpu_core::primitives::machine_type::MachineType;
-use itertools::Itertools;
-use riscv_transpiler::jit::{CounterType, MachineCounters, MAX_NUM_COUNTERS};
-use riscv_transpiler::vm::{
-    Counters, DelegationsAndFamiliesCounters, DelegationsAndUnifiedCounters,
-};
-use riscv_transpiler::witness::delegation::bigint::BigintDelegationWitness;
-use riscv_transpiler::witness::delegation::blake2_g_function::Blake2sGFunctionDelegationWitness;
-use riscv_transpiler::witness::delegation::blake2_round_function::Blake2sRoundFunctionDelegationWitness;
-use riscv_transpiler::witness::delegation::keccak_special5::KeccakSpecial5DelegationWitness;
-use riscv_transpiler::witness::{
-    MemoryOpcodeTracingDataWithTimestamp, NonMemoryOpcodeTracingDataWithTimestamp,
-    UnifiedOpcodeTracingDataWithTimestamp, WitnessTracer,
-};
-use std::any;
-use std::cmp::min;
-use std::collections::{BTreeSet, VecDeque};
-use std::mem::{replace, transmute};
-use std::sync::Arc;
-
-const ADD_SUB_FAMILY_IDX: u8 = CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-    UnrolledNonMemoryCircuitType::AddSubLuiAuipcMop,
-))
-.get_family_idx();
-const JUMP_BRANCH_FAMILY_IDX: u8 = CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-    UnrolledNonMemoryCircuitType::JumpBranchSlt,
-))
-.get_family_idx();
-const SHIFT_BINARY_FAMILY_IDX: u8 = CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-    UnrolledNonMemoryCircuitType::ShiftBinaryCsr,
-))
-.get_family_idx();
-const MUL_DIV_FAMILY_IDX: u8 = CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-    UnrolledNonMemoryCircuitType::MulDivUnsigned,
-))
-.get_family_idx();
-const LOAD_STORE_SUBWORD_FAMILY_IDX: u8 = CircuitType::Unrolled(UnrolledCircuitType::Memory(
-    UnrolledMemoryCircuitType::LoadStoreSubwordOnly,
-))
-.get_family_idx();
-const LOAD_STORE_WORD_FAMILY_IDX: u8 = CircuitType::Unrolled(UnrolledCircuitType::Memory(
-    UnrolledMemoryCircuitType::LoadStoreWordOnly,
-))
-.get_family_idx();
-const BLAKE_DELEGATION_TYPE_ID: u16 =
-    DelegationCircuitType::Blake2WithCompression.get_delegation_type_id();
-const BIGINT_DELEGATION_TYPE_ID: u16 =
-    DelegationCircuitType::BigIntWithControl.get_delegation_type_id();
-const KECCAK_DELEGATION_TYPE_ID: u16 =
-    DelegationCircuitType::KeccakSpecial5.get_delegation_type_id();
-const BLAKE_G_FUNCTION_DELEGATION_TYPE_ID: u16 =
-    DelegationCircuitType::Blake2GFunction.get_delegation_type_id();
-
-pub(crate) trait TracingType {
-    const IS_SPLIT: bool;
-    type Ranges: DataTraceRanges;
-    type Producers: TracingDataProducers<Ranges = Self::Ranges>;
-    type Tracer: Tracer<Ranges = Self::Ranges>;
-    type Counters: Counters + From<MachineCounters>;
-}
-
-pub(crate) struct SplitTracingType;
-
-impl TracingType for SplitTracingType {
-    const IS_SPLIT: bool = true;
-    type Ranges = SplitDataTraceRanges;
-    type Producers = SplitTracingDataProducers;
-    type Tracer = SplitTracer;
-    type Counters = DelegationsAndFamiliesCounters;
-}
-
-pub(crate) struct UnifiedTracingType;
-
-impl TracingType for UnifiedTracingType {
-    const IS_SPLIT: bool = false;
-    type Ranges = UnifiedDataTraceRanges;
-    type Producers = UnifiedTracingDataProducers;
-    type Tracer = UnifiedTracer;
-    type Counters = DelegationsAndUnifiedCounters;
-}
-
-pub(crate) struct PtrRange<T> {
-    pub start: *mut T,
-    pub end: *mut T,
-    pub _chunk: Option<Arc<Vec<T, A>>>,
-}
-
-impl<T> Default for PtrRange<T> {
-    fn default() -> Self {
-        Self {
-            start: std::ptr::null_mut(),
-            end: std::ptr::null_mut(),
-            _chunk: None,
-        }
-    }
-}
-
-// SAFETY: `PtrRange<T>` is a raw pointer pair into a host buffer owned by
-// `_chunk`. The owning chunk is moved across threads alongside the range,
-// so the lifetime of the pointed-to memory matches the receiver thread's
-// access window. Concurrent mutation is controlled by the queue's pop/push
-// discipline, not by the type itself.
-unsafe impl<T> Send for PtrRange<T> {}
-
-pub(crate) trait DataTraceRanges {}
-
-#[derive(Default)]
-pub(crate) struct SplitDataTraceRanges {
-    pub blake_calls: VecDeque<PtrRange<Blake2sRoundFunctionDelegationWitness>>,
-    pub bigint_calls: VecDeque<PtrRange<BigintDelegationWitness>>,
-    pub keccak_calls: VecDeque<PtrRange<KeccakSpecial5DelegationWitness>>,
-    pub blake_g_function_calls: VecDeque<PtrRange<Blake2sGFunctionDelegationWitness>>,
-    pub add_sub_family: VecDeque<PtrRange<NonMemoryOpcodeTracingDataWithTimestamp>>,
-    pub binary_shift_csr_family: VecDeque<PtrRange<NonMemoryOpcodeTracingDataWithTimestamp>>,
-    pub slt_branch_family: VecDeque<PtrRange<NonMemoryOpcodeTracingDataWithTimestamp>>,
-    pub mul_div_family: VecDeque<PtrRange<NonMemoryOpcodeTracingDataWithTimestamp>>,
-    pub word_size_mem_family: VecDeque<PtrRange<MemoryOpcodeTracingDataWithTimestamp>>,
-    pub subword_size_mem_family: VecDeque<PtrRange<MemoryOpcodeTracingDataWithTimestamp>>,
-}
-
-impl DataTraceRanges for SplitDataTraceRanges {}
-
-#[derive(Default)]
-pub(crate) struct UnifiedDataTraceRanges {
-    pub blake_calls: VecDeque<PtrRange<Blake2sRoundFunctionDelegationWitness>>,
-    pub bigint_calls: VecDeque<PtrRange<BigintDelegationWitness>>,
-    pub keccak_calls: VecDeque<PtrRange<KeccakSpecial5DelegationWitness>>,
-    pub blake_g_function_calls: VecDeque<PtrRange<Blake2sGFunctionDelegationWitness>>,
-    pub cycles: VecDeque<PtrRange<UnifiedOpcodeTracingDataWithTimestamp>>,
-}
-
-impl DataTraceRanges for UnifiedDataTraceRanges {}
-
-struct TracerRanges<T: Copy + 'static> {
-    queue: VecDeque<PtrRange<T>>,
-    current: PtrRange<T>,
-    count: usize,
-}
-
-impl<T: Copy + 'static> TracerRanges<T> {
-    fn new(queue: VecDeque<PtrRange<T>>) -> Self {
-        Self {
-            queue,
-            current: PtrRange::default(),
-            count: 0,
-        }
-    }
-
-    /// # Safety
-    ///
-    /// Caller must ensure the producer reserved enough capacity in `queue`
-    /// for every `write` made during a snapshot (see `process_snapshot`).
-    #[inline(always)]
-    unsafe fn write(&mut self, value: T) {
-        self.write_type_unchecked(value);
-    }
-
-    /// # Safety
-    ///
-    /// Same capacity contract as [`Self::write`], plus `U` must equal `T`
-    /// (debug-asserted below). Used to forward dispatch-erased values from
-    /// callers that have a const-generic guarantee of type equality.
-    #[inline(always)]
-    unsafe fn write_type_unchecked<U: Copy + 'static>(&mut self, value: U) {
-        debug_assert_eq!(any::TypeId::of::<T>(), any::TypeId::of::<U>());
-        if core::hint::unlikely(self.current.start == self.current.end) {
-            // SAFETY: capacity precondition: producer pre-queued enough
-            // `PtrRange`s for this snapshot's writes.
-            self.current = self.queue.pop_front().unwrap_unchecked();
-        }
-        // SAFETY: `self.current.start` points into a live `PtrRange` reserved
-        // for this tracer; `T == U` debug-checked above.
-        *(self.current.start as *mut U) = value;
-        self.current.start = self.current.start.add(1);
-        self.count += 1;
-    }
-}
-
-pub(crate) trait Tracer: WitnessTracer {
-    type Ranges: DataTraceRanges + Send;
-
-    fn new(trace_ranges: Self::Ranges) -> Self;
-}
-
-pub(crate) struct SplitTracer {
-    blake_calls: TracerRanges<Blake2sRoundFunctionDelegationWitness>,
-    bigint_calls: TracerRanges<BigintDelegationWitness>,
-    keccak_calls: TracerRanges<KeccakSpecial5DelegationWitness>,
-    blake_g_function_calls: TracerRanges<Blake2sGFunctionDelegationWitness>,
-    add_sub_family: TracerRanges<NonMemoryOpcodeTracingDataWithTimestamp>,
-    binary_shift_csr_family: TracerRanges<NonMemoryOpcodeTracingDataWithTimestamp>,
-    slt_branch_family: TracerRanges<NonMemoryOpcodeTracingDataWithTimestamp>,
-    mul_div_family: TracerRanges<NonMemoryOpcodeTracingDataWithTimestamp>,
-    word_size_mem_family: TracerRanges<MemoryOpcodeTracingDataWithTimestamp>,
-    subword_size_mem_family: TracerRanges<MemoryOpcodeTracingDataWithTimestamp>,
-}
-
-impl Tracer for SplitTracer {
-    type Ranges = SplitDataTraceRanges;
-
-    fn new(trace_ranges: Self::Ranges) -> Self {
-        Self {
-            blake_calls: TracerRanges::new(trace_ranges.blake_calls),
-            bigint_calls: TracerRanges::new(trace_ranges.bigint_calls),
-            keccak_calls: TracerRanges::new(trace_ranges.keccak_calls),
-            blake_g_function_calls: TracerRanges::new(trace_ranges.blake_g_function_calls),
-            add_sub_family: TracerRanges::new(trace_ranges.add_sub_family),
-            binary_shift_csr_family: TracerRanges::new(trace_ranges.binary_shift_csr_family),
-            slt_branch_family: TracerRanges::new(trace_ranges.slt_branch_family),
-            mul_div_family: TracerRanges::new(trace_ranges.mul_div_family),
-            word_size_mem_family: TracerRanges::new(trace_ranges.word_size_mem_family),
-            subword_size_mem_family: TracerRanges::new(trace_ranges.subword_size_mem_family),
-        }
-    }
-}
-
-// `SplitTracer` and `UnifiedTracer` both name their four delegation
-// `TracerRanges` fields identically (`blake_calls` / `bigint_calls` /
-// `keccak_calls` / `blake_g_function_calls`), so `WitnessTracer::write_delegation`
-// dispatches identically for both — this macro is the shared method body,
-// invoked once per `impl WitnessTracer for { Split, Unified }Tracer` block.
-macro_rules! impl_write_delegation {
-    () => {
-        #[inline(always)]
-        fn write_delegation<
-            const DELEGATION_TYPE: u16,
-            const REG_ACCESSES: usize,
-            const INDIRECT_READS: usize,
-            const INDIRECT_WRITES: usize,
-            const VARIABLE_OFFSETS: usize,
-        >(
-            &mut self,
-            data: riscv_transpiler::witness::DelegationWitness<
-                REG_ACCESSES,
-                INDIRECT_READS,
-                INDIRECT_WRITES,
-                VARIABLE_OFFSETS,
-            >,
-        ) {
-            // SAFETY: `unreachable_unchecked` covers the upstream
-            // `WitnessTracer::write_delegation` contract (DELEGATION_TYPE outside
-            // the enumerated set). `write_type_unchecked` is sound because each
-            // arm pairs a const DELEGATION_TYPE with the matching witness type T
-            // by construction.
-            unsafe {
-                if const { DELEGATION_TYPE == BLAKE_DELEGATION_TYPE_ID } {
-                    self.blake_calls.write_type_unchecked(data)
-                } else if const { DELEGATION_TYPE == BIGINT_DELEGATION_TYPE_ID } {
-                    self.bigint_calls.write_type_unchecked(data)
-                } else if const { DELEGATION_TYPE == KECCAK_DELEGATION_TYPE_ID } {
-                    self.keccak_calls.write_type_unchecked(data)
-                } else if const { DELEGATION_TYPE == BLAKE_G_FUNCTION_DELEGATION_TYPE_ID } {
-                    self.blake_g_function_calls.write_type_unchecked(data)
-                } else {
-                    core::hint::unreachable_unchecked()
-                };
-            }
-        }
-    };
-}
-
-impl WitnessTracer for SplitTracer {
-    #[inline(always)]
-    fn needs_tracing_data_for_circuit_family<const FAMILY: u8>(&self) -> bool {
-        true
-    }
-
-    #[inline(always)]
-    fn needs_tracing_data_for_delegation_type<const DELEGATION_TYPE: u16>(&self) -> bool {
-        true
-    }
-
-    #[inline(always)]
-    fn write_non_memory_family_data<const FAMILY: u8>(
-        &mut self,
-        data: NonMemoryOpcodeTracingDataWithTimestamp,
-    ) {
-        // SAFETY: `unreachable_unchecked` is reached only if the upstream
-        // `WitnessTracer` contract is violated (FAMILY outside the enumerated
-        // set). The `.write()` calls rely on capacity pre-queued by the
-        // matching family `Producer` for this snapshot.
-        unsafe {
-            if const { FAMILY == ADD_SUB_FAMILY_IDX } {
-                self.add_sub_family.write(data)
-            } else if const { FAMILY == JUMP_BRANCH_FAMILY_IDX } {
-                self.slt_branch_family.write(data)
-            } else if const { FAMILY == SHIFT_BINARY_FAMILY_IDX } {
-                self.binary_shift_csr_family.write(data)
-            } else if const { FAMILY == MUL_DIV_FAMILY_IDX } {
-                self.mul_div_family.write(data)
-            } else {
-                core::hint::unreachable_unchecked()
-            };
-        }
-    }
-
-    #[inline(always)]
-    fn write_memory_family_data<const FAMILY: u8>(
-        &mut self,
-        data: MemoryOpcodeTracingDataWithTimestamp,
-    ) {
-        // SAFETY: as in `write_non_memory_family_data` — `unreachable_unchecked`
-        // covers the upstream FAMILY contract; `.write()` relies on pre-queued
-        // capacity owned by the matching family `Producer`.
-        unsafe {
-            if const { FAMILY == LOAD_STORE_SUBWORD_FAMILY_IDX } {
-                self.subword_size_mem_family.write(data)
-            } else if const { FAMILY == LOAD_STORE_WORD_FAMILY_IDX } {
-                self.word_size_mem_family.write(data)
-            } else {
-                core::hint::unreachable_unchecked()
-            };
-        }
-    }
-
-    impl_write_delegation!();
-}
-
-pub(crate) struct UnifiedTracer {
-    blake_calls: TracerRanges<Blake2sRoundFunctionDelegationWitness>,
-    bigint_calls: TracerRanges<BigintDelegationWitness>,
-    keccak_calls: TracerRanges<KeccakSpecial5DelegationWitness>,
-    blake_g_function_calls: TracerRanges<Blake2sGFunctionDelegationWitness>,
-    cycles: TracerRanges<UnifiedOpcodeTracingDataWithTimestamp>,
-}
-
-impl Tracer for UnifiedTracer {
-    type Ranges = UnifiedDataTraceRanges;
-
-    fn new(trace_ranges: Self::Ranges) -> Self {
-        Self {
-            blake_calls: TracerRanges::new(trace_ranges.blake_calls),
-            bigint_calls: TracerRanges::new(trace_ranges.bigint_calls),
-            keccak_calls: TracerRanges::new(trace_ranges.keccak_calls),
-            blake_g_function_calls: TracerRanges::new(trace_ranges.blake_g_function_calls),
-            cycles: TracerRanges::new(trace_ranges.cycles),
-        }
-    }
-}
-
-impl WitnessTracer for UnifiedTracer {
-    #[inline(always)]
-    fn needs_tracing_data_for_circuit_family<const FAMILY: u8>(&self) -> bool {
-        true
-    }
-
-    #[inline(always)]
-    fn needs_tracing_data_for_delegation_type<const DELEGATION_TYPE: u16>(&self) -> bool {
-        true
-    }
-
-    #[inline(always)]
-    fn write_non_memory_family_data<const FAMILY: u8>(
-        &mut self,
-        data: NonMemoryOpcodeTracingDataWithTimestamp,
-    ) {
-        // SAFETY: capacity pre-queued by the unified `Producer` for the
-        // `cycles` stream during `process_snapshot`.
-        unsafe {
-            self.cycles
-                .write(UnifiedOpcodeTracingDataWithTimestamp::NonMem(data))
-        }
-    }
-
-    #[inline(always)]
-    fn write_memory_family_data<const FAMILY: u8>(
-        &mut self,
-        data: MemoryOpcodeTracingDataWithTimestamp,
-    ) {
-        // SAFETY: as above — capacity pre-queued by the unified `Producer`.
-        unsafe {
-            self.cycles
-                .write(UnifiedOpcodeTracingDataWithTimestamp::Mem(data))
-        }
-    }
-
-    impl_write_delegation!();
-}
-
-trait TracingDataProducerType: Sized {
-    fn produce_tracing_data(holder: ChunkedTraceHolder<Self, A>) -> TracingDataHost<A>;
-}
-
-// `DelegationTracingDataHostSource` is foreign (defined in `gpu_circuit_prover`),
-// so a blanket `impl<T: DelegationTracingDataHostSource>` is not coherent here —
-// the compiler cannot prove the unrolled types below don't also implement it.
-// Enumerate the four delegation witness types instead (the trait has exactly
-// these four impls upstream).
-macro_rules! impl_delegation_tracing_data_producer {
-    ($($ty:ty),+ $(,)?) => {$(
-        impl TracingDataProducerType for $ty {
-            fn produce_tracing_data(holder: ChunkedTraceHolder<Self, A>) -> TracingDataHost<A> {
-                TracingDataHost::Delegation(<Self as DelegationTracingDataHostSource>::get(holder))
-            }
-        }
-    )+};
-}
-impl_delegation_tracing_data_producer!(
-    BigintDelegationWitness,
-    Blake2sRoundFunctionDelegationWitness,
-    Blake2sGFunctionDelegationWitness,
-    KeccakSpecial5DelegationWitness,
-);
-
-impl TracingDataProducerType for MemoryOpcodeTracingDataWithTimestamp {
-    fn produce_tracing_data(holder: ChunkedTraceHolder<Self, A>) -> TracingDataHost<A> {
-        TracingDataHost::Unrolled(UnrolledTracingDataHost::Memory(holder))
-    }
-}
-
-impl TracingDataProducerType for NonMemoryOpcodeTracingDataWithTimestamp {
-    fn produce_tracing_data(holder: ChunkedTraceHolder<Self, A>) -> TracingDataHost<A> {
-        TracingDataHost::Unrolled(UnrolledTracingDataHost::NonMemory(holder))
-    }
-}
-
-impl TracingDataProducerType for UnifiedOpcodeTracingDataWithTimestamp {
-    fn produce_tracing_data(holder: ChunkedTraceHolder<Self, A>) -> TracingDataHost<A> {
-        TracingDataHost::Unrolled(UnrolledTracingDataHost::Unified(holder))
-    }
-}
-
-/// Per-circuit trace-row count used to slice cycle counters into circuit-sized
-/// partitions. The full `2^N` rows are usable (no reserved padding row).
-fn cycles_per_circuit_for(circuit_type: CircuitType) -> usize {
-    circuit_type.get_domain_size()
-}
-
-struct TracingDataProducer<T: TracingDataProducerType> {
-    circuit_type: CircuitType,
-    cycles_per_circuit: usize,
-    free_allocators: Receiver<A>,
-    results: Sender<WorkerResult<A>>,
-    current_circuit_index: usize,
-    chunks: VecDeque<Arc<Vec<T, A>>>,
-    participating_snapshot_indexes: BTreeSet<usize>,
-}
-
-impl<T: TracingDataProducerType> TracingDataProducer<T> {
-    pub fn new(
-        circuit_type: CircuitType,
-        free_allocators: Receiver<A>,
-        results: Sender<WorkerResult<A>>,
-    ) -> Self {
-        Self {
-            circuit_type,
-            cycles_per_circuit: cycles_per_circuit_for(circuit_type),
-            free_allocators,
-            results,
-            current_circuit_index: 0,
-            chunks: VecDeque::new(),
-            participating_snapshot_indexes: BTreeSet::new(),
-        }
-    }
-
-    pub fn process_snapshot(
-        &mut self,
-        snapshot_index: usize,
-        mut start: usize,
-        end: usize,
-        trace_ranges: &mut VecDeque<PtrRange<T>>,
-    ) {
-        while start != end {
-            let cycles_per_circuit = self.cycles_per_circuit;
-            let next_circuit_boundary = (start + 1).next_multiple_of(cycles_per_circuit);
-            let next_circuit_index = next_circuit_boundary / cycles_per_circuit;
-            assert_eq!(next_circuit_index, self.current_circuit_index + 1);
-            if self.chunks.back().map_or(true, |v| v.len() == v.capacity()) {
-                let allocator = self
-                    .free_allocators
-                    .recv()
-                    .expect("tracing allocator channel closed while growing a trace chunk");
-                let capacity = allocator.capacity() / size_of::<T>();
-                let chunk = Arc::new(Vec::with_capacity_in(capacity, allocator));
-                self.chunks.push_back(chunk)
-            };
-            let chunk = self.chunks.back_mut().unwrap();
-            // SAFETY: prior `Arc::clone`s of `chunk` live inside `PtrRange`s
-            // queued for the tracer/consumer threads, but those threads run
-            // strictly after this producer finishes the snapshot, so no
-            // concurrent access to the `Vec` exists right now.
-            let chunk_mut = unsafe { Arc::get_mut_unchecked(chunk) };
-            let spare_capacity = chunk_mut.spare_capacity_mut();
-            let end = min(end, next_circuit_boundary);
-            let diff = min(spare_capacity.len(), end - start);
-            assert_ne!(diff, 0);
-            // SAFETY: `start_ptr..end_ptr` is `diff` elements inside
-            // `spare_capacity` (bounded by `diff <= spare_capacity.len()`).
-            // `set_len` extends the Vec over memory the tracer fills via the
-            // raw `PtrRange` before any reader sees the chunk via Vec
-            // accessors; `T: Copy + 'static` so no `Drop` runs on slots that
-            // remain unwritten until then.
-            let ptr_range = unsafe {
-                let start_ptr = spare_capacity.as_mut_ptr() as *mut T;
-                let end_ptr = start_ptr.add(diff);
-                chunk_mut.set_len(chunk_mut.len() + diff);
-                PtrRange {
-                    start: start_ptr,
-                    end: end_ptr,
-                    _chunk: Some(chunk.clone()),
-                }
-            };
-            trace_ranges.push_back(ptr_range);
-            self.participating_snapshot_indexes.insert(snapshot_index);
-            start += diff;
-            if start.is_multiple_of(cycles_per_circuit) {
-                assert_eq!(start / cycles_per_circuit, next_circuit_index);
-                self.produce_and_send_result();
-                self.current_circuit_index = next_circuit_index;
-            }
-        }
-    }
-
-    fn produce_and_send_result(&mut self) {
-        let chunks = self.chunks.drain(..).collect_vec();
-        let holder = ChunkedTraceHolder { chunks };
-        let tracing_data = T::produce_tracing_data(holder);
-        let participating_snapshot_indexes =
-            replace(&mut self.participating_snapshot_indexes, BTreeSet::new());
-        let data = TracingData {
-            circuit_type: self.circuit_type,
-            sequence_id: self.current_circuit_index,
-            tracing_data,
-            participating_snapshot_indexes,
-        };
-        let result = WorkerResult::TracingData(data);
-        self.results
-            .send(result)
-            .expect("tracing results channel closed while sending tracing data");
-    }
-
-    pub fn finalize(mut self) {
-        if !self.chunks.is_empty() {
-            self.produce_and_send_result()
-        }
-    }
-}
-
-pub(crate) trait TracingDataProducers {
-    type Ranges: DataTraceRanges + Send;
-
-    fn new(
-        machine_type: MachineType,
-        free_allocators: Receiver<A>,
-        results: Sender<WorkerResult<A>>,
-    ) -> Self;
-
-    fn process_snapshot(
-        &mut self,
-        snapshot_index: usize,
-        initial_counters: &[u64; MAX_NUM_COUNTERS],
-        final_counters: &[u64; MAX_NUM_COUNTERS],
-    ) -> Self::Ranges;
-
-    fn finalize(self);
-}
-
-/// The four delegation producers (blake / bigint / keccak /
-/// blake_g_function), constructed and torn down identically by
-/// `SplitTracingDataProducers` and `UnifiedTracingDataProducers` — the two
-/// differ only in which additional (per-family or unified-cycle) producers
-/// accompany this shared set.
-struct DelegationProducers {
-    blake_producer: TracingDataProducer<Blake2sRoundFunctionDelegationWitness>,
-    bigint_producer: TracingDataProducer<BigintDelegationWitness>,
-    keccak_producer: TracingDataProducer<KeccakSpecial5DelegationWitness>,
-    blake_g_function_producer: TracingDataProducer<Blake2sGFunctionDelegationWitness>,
-}
-
-impl DelegationProducers {
-    fn new(free_allocators: &Receiver<A>, results: &Sender<WorkerResult<A>>) -> Self {
-        let blake_producer = TracingDataProducer::<Blake2sRoundFunctionDelegationWitness>::new(
-            CircuitType::Delegation(DelegationCircuitType::Blake2WithCompression),
-            free_allocators.clone(),
-            results.clone(),
-        );
-        let bigint_producer = TracingDataProducer::<BigintDelegationWitness>::new(
-            CircuitType::Delegation(DelegationCircuitType::BigIntWithControl),
-            free_allocators.clone(),
-            results.clone(),
-        );
-        let keccak_producer = TracingDataProducer::<KeccakSpecial5DelegationWitness>::new(
-            CircuitType::Delegation(DelegationCircuitType::KeccakSpecial5),
-            free_allocators.clone(),
-            results.clone(),
-        );
-        let blake_g_function_producer =
-            TracingDataProducer::<Blake2sGFunctionDelegationWitness>::new(
-                CircuitType::Delegation(DelegationCircuitType::Blake2GFunction),
-                free_allocators.clone(),
-                results.clone(),
-            );
-        Self {
-            blake_producer,
-            bigint_producer,
-            keccak_producer,
-            blake_g_function_producer,
-        }
-    }
-
-    fn finalize(self) {
-        self.blake_producer.finalize();
-        self.bigint_producer.finalize();
-        self.keccak_producer.finalize();
-        self.blake_g_function_producer.finalize();
-    }
-}
-
-pub(crate) struct SplitTracingDataProducers {
-    delegation: DelegationProducers,
-    add_sub_family_producer: TracingDataProducer<NonMemoryOpcodeTracingDataWithTimestamp>,
-    binary_shift_csr_family_producer: TracingDataProducer<NonMemoryOpcodeTracingDataWithTimestamp>,
-    slt_branch_family_producer: TracingDataProducer<NonMemoryOpcodeTracingDataWithTimestamp>,
-    mul_div_family_producer: TracingDataProducer<NonMemoryOpcodeTracingDataWithTimestamp>,
-    word_size_mem_family_producer: TracingDataProducer<MemoryOpcodeTracingDataWithTimestamp>,
-    subword_size_mem_family_producer: TracingDataProducer<MemoryOpcodeTracingDataWithTimestamp>,
-}
-
-impl TracingDataProducers for SplitTracingDataProducers {
-    type Ranges = SplitDataTraceRanges;
-
-    fn new(
-        _machine_type: MachineType,
-        free_allocators: Receiver<A>,
-        results: Sender<WorkerResult<A>>,
-    ) -> Self {
-        let delegation = DelegationProducers::new(&free_allocators, &results);
-        let add_sub_family_producer =
-            TracingDataProducer::<NonMemoryOpcodeTracingDataWithTimestamp>::new(
-                CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-                    UnrolledNonMemoryCircuitType::AddSubLuiAuipcMop,
-                )),
-                free_allocators.clone(),
-                results.clone(),
-            );
-        let binary_shift_csr_family_producer =
-            TracingDataProducer::<NonMemoryOpcodeTracingDataWithTimestamp>::new(
-                CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-                    UnrolledNonMemoryCircuitType::ShiftBinaryCsr,
-                )),
-                free_allocators.clone(),
-                results.clone(),
-            );
-        let slt_branch_family_producer =
-            TracingDataProducer::<NonMemoryOpcodeTracingDataWithTimestamp>::new(
-                CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-                    UnrolledNonMemoryCircuitType::JumpBranchSlt,
-                )),
-                free_allocators.clone(),
-                results.clone(),
-            );
-        let mul_div_family_producer =
-            TracingDataProducer::<NonMemoryOpcodeTracingDataWithTimestamp>::new(
-                CircuitType::Unrolled(UnrolledCircuitType::NonMemory(
-                    UnrolledNonMemoryCircuitType::MulDivUnsigned,
-                )),
-                free_allocators.clone(),
-                results.clone(),
-            );
-        let word_size_mem_family_producer =
-            TracingDataProducer::<MemoryOpcodeTracingDataWithTimestamp>::new(
-                CircuitType::Unrolled(UnrolledCircuitType::Memory(
-                    UnrolledMemoryCircuitType::LoadStoreWordOnly,
-                )),
-                free_allocators.clone(),
-                results.clone(),
-            );
-        let subword_size_mem_family_producer =
-            TracingDataProducer::<MemoryOpcodeTracingDataWithTimestamp>::new(
-                CircuitType::Unrolled(UnrolledCircuitType::Memory(
-                    UnrolledMemoryCircuitType::LoadStoreSubwordOnly,
-                )),
-                free_allocators,
-                results,
-            );
-        Self {
-            delegation,
-            add_sub_family_producer,
-            binary_shift_csr_family_producer,
-            slt_branch_family_producer,
-            mul_div_family_producer,
-            word_size_mem_family_producer,
-            subword_size_mem_family_producer,
-        }
-    }
-
-    fn process_snapshot(
-        &mut self,
-        snapshot_index: usize,
-        initial_counters: &[u64; MAX_NUM_COUNTERS],
-        final_counters: &[u64; MAX_NUM_COUNTERS],
-    ) -> Self::Ranges {
-        let mut trace_ranges = SplitDataTraceRanges::default();
-        for i in 0..CounterType::FormalEnd as u8 {
-            // SAFETY: loop bound `0..CounterType::FormalEnd as u8` keeps `i`
-            // within the enum's defined `#[repr(u8)]` discriminants.
-            let counter_type = unsafe { transmute(i) };
-            let index = i as usize;
-            let initial_count = initial_counters[index] as usize;
-            let final_count = final_counters[index] as usize;
-            match counter_type {
-                CounterType::AddSubLui => self.add_sub_family_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.add_sub_family,
-                ),
-                CounterType::BranchSlt => self.slt_branch_family_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.slt_branch_family,
-                ),
-                CounterType::ShiftBinary => self.binary_shift_csr_family_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.binary_shift_csr_family,
-                ),
-                CounterType::MulDiv => self.mul_div_family_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.mul_div_family,
-                ),
-                CounterType::MemWord => self.word_size_mem_family_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.word_size_mem_family,
-                ),
-                CounterType::MemSubword => self.subword_size_mem_family_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.subword_size_mem_family,
-                ),
-                CounterType::BlakeDelegation => self.delegation.blake_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.blake_calls,
-                ),
-                CounterType::BigintDelegation => self.delegation.bigint_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.bigint_calls,
-                ),
-                CounterType::KeccakDelegation => self.delegation.keccak_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.keccak_calls,
-                ),
-                CounterType::BlakeGFunctionDelegation => {
-                    self.delegation.blake_g_function_producer.process_snapshot(
-                        snapshot_index,
-                        initial_count,
-                        final_count,
-                        &mut trace_ranges.blake_g_function_calls,
-                    )
-                }
-                _ => unreachable!(),
-            }
-        }
-        trace_ranges
-    }
-
-    fn finalize(self) {
-        self.delegation.finalize();
-        self.add_sub_family_producer.finalize();
-        self.binary_shift_csr_family_producer.finalize();
-        self.slt_branch_family_producer.finalize();
-        self.mul_div_family_producer.finalize();
-        self.word_size_mem_family_producer.finalize();
-        self.subword_size_mem_family_producer.finalize();
-    }
-}
-
-pub(crate) struct UnifiedTracingDataProducers {
-    delegation: DelegationProducers,
-    cycles_producer: TracingDataProducer<UnifiedOpcodeTracingDataWithTimestamp>,
-}
-
-impl TracingDataProducers for UnifiedTracingDataProducers {
-    type Ranges = UnifiedDataTraceRanges;
-
-    fn new(
-        machine_type: MachineType,
-        free_allocators: Receiver<A>,
-        results: Sender<WorkerResult<A>>,
-    ) -> Self {
-        assert_eq!(machine_type, MachineType::Reduced);
-        let delegation = DelegationProducers::new(&free_allocators, &results);
-        let cycles_producer = TracingDataProducer::<UnifiedOpcodeTracingDataWithTimestamp>::new(
-            CircuitType::Unrolled(UnrolledCircuitType::Unified),
-            free_allocators.clone(),
-            results.clone(),
-        );
-        Self {
-            delegation,
-            cycles_producer,
-        }
-    }
-
-    fn process_snapshot(
-        &mut self,
-        snapshot_index: usize,
-        initial_counters: &[u64; MAX_NUM_COUNTERS],
-        final_counters: &[u64; MAX_NUM_COUNTERS],
-    ) -> Self::Ranges {
-        let mut trace_ranges = UnifiedDataTraceRanges::default();
-        let mut cycles_initial_count = 0;
-        let mut cycles_final_count = 0;
-        for i in 0..CounterType::FormalEnd as u8 {
-            // SAFETY: loop bound `0..CounterType::FormalEnd as u8` keeps `i`
-            // within the enum's defined `#[repr(u8)]` discriminants.
-            let counter_type = unsafe { transmute(i) };
-            let index = i as usize;
-            let initial_count = initial_counters[index] as usize;
-            let final_count = final_counters[index] as usize;
-            match counter_type {
-                CounterType::AddSubLui
-                | CounterType::BranchSlt
-                | CounterType::ShiftBinary
-                | CounterType::MulDiv
-                | CounterType::MemWord
-                | CounterType::MemSubword => {
-                    cycles_initial_count += initial_count;
-                    cycles_final_count += final_count;
-                }
-                CounterType::BlakeDelegation => self.delegation.blake_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.blake_calls,
-                ),
-                CounterType::BigintDelegation => self.delegation.bigint_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.bigint_calls,
-                ),
-                CounterType::KeccakDelegation => self.delegation.keccak_producer.process_snapshot(
-                    snapshot_index,
-                    initial_count,
-                    final_count,
-                    &mut trace_ranges.keccak_calls,
-                ),
-                CounterType::BlakeGFunctionDelegation => {
-                    self.delegation.blake_g_function_producer.process_snapshot(
-                        snapshot_index,
-                        initial_count,
-                        final_count,
-                        &mut trace_ranges.blake_g_function_calls,
-                    )
-                }
-                _ => unreachable!(),
-            }
-        }
-        self.cycles_producer.process_snapshot(
-            snapshot_index,
-            cycles_initial_count,
-            cycles_final_count,
-            &mut trace_ranges.cycles,
-        );
-        trace_ranges
-    }
-
-    fn finalize(self) {
-        self.delegation.finalize();
-        self.cycles_producer.finalize();
-    }
-}
